Refactor build script into modular package
